moved functions in crud.py to db.py
